Calendrical algebra for converting between time units

This S7 generic function defines the calendrical relationships between two time units, and is the building block for defining calendars in mixtime. It calculates how many x time units fit into one y unit. Some cyclical granularities are context-dependent (such as the number of days in a month), and so an optional time point can be provided with at.

calendar_algebra(x, y, ...)

Arguments

x

The primary time unit

y

The time unit to convert x into

at

Optional time point for context-dependent ratios

Value

Numeric describing how many x time units fit into y at time at.

Details

The methods are dispatched based on the shortest path along defined methods. This allows for defining only the direct relationships between adjacent time units, and relying on graph traversal to find how to convert between more distant units. For example the number of seconds in an hour can be calculated from the number of seconds in a minute and then number of minutes in an hour.

If a method is defined for converting between time units of different calendar systems (e.g., Gregorian calendar days to Chinese calendar days), then that method can be used to convert times at any granularity between the two systems.

Examples

# There are 12 months in a year
calendar_algebra(tu_year(1L), tu_month(1L))
#> [1] 12

# There are 7 days in a week
calendar_algebra(tu_week(1L), tu_day(1L))
#> [1] 7

# There are 3600 seconds in an hour
calendar_algebra(tu_hour(1L), tu_second(1L))
#> Error in if (start == end) {    return(start)}: missing value where TRUE/FALSE needed

# There are 18 "2 months" in 3 years
calendar_algebra(tu_year(3L), tu_month(2L))
#> [1] 18